moving foreman's installer code to a library that can be referenced
This allows Katello and others to re-use the code in Foreman's kafo
launcher.  Originally I had thought we could move this into kafo
directly but almost *all* of the code in the old bin/foreman-installer
and now lib/foreman_installer/launcher.rb was specific to Foreman and
not really well suited to Kafo directly.