fixes #12498 - adding initial config and answers for capsule scenario.
This commit contains the initial answer and configuration file to support
the 'capsule' as a kafo installer scenario on top of the foreman-proxy.
As a result, the user could use the 'foreman-installer' to install
a capsule vs the current 'capsule-installer'.

Note: this is one of the changes being put in place to support the
'Katello on Foreman' feature.