fixes #14453 - build and install Kafo parser cache
The parser cache will provide support for Puppet in AIO in the near
future with Kafo by not loading Puppet or puppet-strings, and provides
a performance boost to the installer startup. Requires Kafo 0.7.3.

Be careful to preserve mtimes of manifest files to ensure the cache
validity. Extra kafo-export-params search dirs were removed as all
current packages install correctly into /usr/bin or PATH, and it is
now required to build.