Fixes #20282, #14667 - Provide more secure defaults for apache (#236)
This sets the default ciphersuites to the recommended for the time of
writing for the supported servers and browsers. This also disables
TRACE method, which is not a security vulnerability but comes up often
in automated security audits and isn't required for proper functioning
of Foreman.

(cherry picked from commit 267653fadea7f443396ecf1443777dc5f6748ba4)