Add foreman-proxy-certs-generate
This is essentially a separate installer that uses the same modules. It
does read the scenario from the normal installer to determine Foreman's
initial organization for the certificate name. This is implemented as a
hook.

It is separate because then Kafo disables multiple scenario management.

Because the answers are never stored, they don't need migrations and can
be installed to /usr/share/foreman-installer/katello-certs. The separate
subdir is there to prevent sharing of hooks since kafo always uses
$installer_dir/hooks.