Refs #26330 - Do not add Foreman to Puppet in Katello
In the Katello scenario we don't need Foreman to be able to read the
Puppet certs. This is the only reason we add Foreman to the Puppet user
group by default. This change allows users to install on systems without
Puppet server present.

Until Puppet 4 the agent package also ensured the Puppet user group
existed so this wasn't a problem with Puppet 3.