Fixes #26409 - Locking packages after installation
This patch add a new option --[no-]lock-packages-versions.
The option enables/disables version locking of installed packages.
Once used the choice is stored in the scenario "custom" hash and
is reused in the following installer runs.

If the packages are locked when the installer starts
unlock is performed before installer proceed with the installation.