Fixes #27848 - Disable HTTP by default on Katello's proxy
The HTTP and templates features are useful when a host needs to be
provisioned and can't reach the Foreman server directly. Since this is
by definition the Foreman server, the client should be able to reach it.

The values are dropped since they match the module defaults. It doesn't
implement a migration to not touch existing installations which might
have started to rely on it.

This reverts #19720.