Expand the create-migration script
This allows creating multiple migrations at the same time following the
correct directory conventions. This makes it easy to create the same
migration for multiple scenarios at the same time. For this the argument
order needs to be swapped. It also accepts the code on STDIN to make it
a single operation.