Fixes #28331 - Base memory checks on tuning profile
Each tuning profile has different system requirements so a generic 8 GB
memory check is no longer sufficient. This patch always validates the
tuning profile to be valid and uses it to check memory and CPU cores. It
also uses facter rather than trying to determine it manually which makes
the code easier to read.

By always validating the tuning, it becomes important that users are
migrated when a tuning profile is removed.