Fixes #29437 - Drop PuppetDB module
The intent was to make it possible to install PuppetDB but there’s no
documentation, it’s tricky to set up and we don’t verify it (it may
actually be broken). Adding to that, its maintainers are largely
unresponsive. Adding support for new distribution versions and other
module versions is very painful.