Refs #30193 - Correct version locking
6e1a07bc35bfe97927dfac6477a1faf53092c58f moved version locking from
Katello to the common hooks. This was unintentional and should have been
kept for Katello, at least for now.

However, it did try to fix a real issue where yum was locked while the
hooks tried to make changes. However, during pre_validations it may
still find an issue and roll back so no changes should be made. To solve
this, the hook is moved to pre. This is more correct because at this
point Kafo, has committed all the answers and is going to make changes
to the system using Puppet.

(cherry picked from commit 6a5daf8c21902ab44272868e7492a56379d3946a)