Add module checkout scripts
This is mainly intended for the stable branches. It allows checking out
all modules based on Puppetfile.lock. This is done by having a directory
called workdir. Within that there's a directory 'bare' which contains
all the modules as bare checkouts. Then in the modules a directory is
created for the MAJOR.MINOR version. Within that directory there's a
checkout for each module owned by theforeman/Katello.

The output of module-versions is changed to make it easier in the
checkout-module script.