Refs #29799 - Define clear mongo methods in module
3dcd45bef6f56a7af898eddf3283229f0e33d065 moved these methods to the
common context but defined them outside of the module. This means they
were not usable in the pre context. That leads some some infinite loop
until the machine is out of memory.

It also handles the case when katello::pulp_db_seeds returns nil, which
is always the case on a content proxy.