Fixes #31135: Introduce execute! and execute for fatal and non-fatal commands
Introduce two separate functions that run a command. execute! will
run a command and exit if the status indicates failure. execute will
run a command and return the status of the command to let the hook
take further action. execute is intended for actions that require
information to decide what further action to take or if the action
is non-fatal.

This present previously seen errors where detecting foreman-maintain
exists on the system or foreman-maintain checks for a feature
from presenting as an error incorrectly to the user.