Fixes #31326: Error in certs check if using only a shortname
Katello requires an FQDN be used for installation and thus the
same requirement exists for any custom certificates being used.
This adds a check that the CN is not set to a shortname, and/or
if using Subject Alt Name that there is not a single SAN that
is a shortname.