reexec systemd after installing SELinux packages
Starting systemd 245 the daemon reloads SELinux contexts after a
daemon-reload, but prior to that it needs to reexec. Both EL7 and EL8
are too old so this runs reexec. To limit the impact, it only does so if
a package change was detected. This can fail if a metapackage (katello)
already pulled in all packages but there's no reliable way to detect if
a reexec is needed. Always running daemon-reexec can be unexpected.