Fixes #31834 - Use PostgreSQL 12 on EL8
This aligns Red Hat 7 and Red Hat 8 on PostgreSQL 12. It uses the
a specific git commit for DNF module support but without dropping Puppet
5 to allow for a cherry pick to 2.4. After develop drops Puppet 5
support, it should switch to >= 7.0.1 as a version specifyer.

It may break existing installations on PostgreSQL 10 since there is no
upgrade path. Those users are recommended to set version to 10 in their
custom-hiera.yaml or upgrade to PostgreSQL 12.