Update Katello modules for 2.4.0-rc3
This is a much larger update than you would normally want. Sadly, a lot
of these fixes are needed to work.

It also corrects some broken locks that were not in sync with the
Puppetfile. The transition module was needed by pulp but that was
removed in e5281240e8f97d821dda23d271a3b214d107590a. This makes sure
it's no longer present.