Restore DNF module management for PG on EL8
In fdc20de529ebcaba1708bf9afa43a23e96a7858a the version of
puppetlabs/postgresql was pinned to an unreleased version which does
have DNF module management but doesn't drop Puppet 5. Then in
157bd6a7165e11c87e53aa5338b57e58018ed8ee it was pinned to >= 6.10.1
which is too old and lacks DNF module management. That means it ends up
using the default PostgreSQL module which is version 10.

Fixes: 157bd6a7165e11c87e53aa5338b57e58018ed8ee