Fixes #33046: Move scenarios to /var/lib/foreman-installer
This change moves scenarios from system config directory to the
shared state directory in /var/lib. The answers and config files
act as databases recording the current state of a scenario and
are not intended to be directly edited. Direct changes to the answers
file or configuration file can lead to inconsistencies or improper
values are that intended to be caught by validation. Moving to
the shared state directory draws this line better and emphasizes to
users that they should not be touching these files directly.