Fixes #33191 - Debian puppet-agent-oauth reinstall support
This has been supported for a while on RPM-based distros. This adds it
on Debian.

It also changes it to log an error on noop that it detected an incorrect
situation.

Credits to @timdeluxe for the specific dpkg/apt commands. This PR takes
a bit different implementation approach by also looking at facts. This
should save some useless command executions on both platforms.