Refs #33214: Reset puma min threads to default from puppet module
If the default value for foreman_service_puma_threads_min is found,
which is zero, this deletes that value from the answers file. This
allows the new default to be populated from the puppet module. The
new default is to set min threads equal to max threads. If the minimum
is anything other than zero than we can assume the user has customized
the value.