Refs #33277: Unset foreman_service_puma_workers and foreman_service_puma_threads_max if set to default value
This resets the foreman_server_puma_workers value to pick up the
newest default which is undef. The change from https://github.com/theforeman/puppet-foreman/pull/986
makes the default a dynamically calculated value that should
provide better out of the box experience for users.

This resets the value of foreman_service_puma_threads_max to the new
default value coming from puppet-foreman of 5. The original default of
16 has been determined to likely be too high for Ruby MRI based installations
and resetting to the recommended default by Puma should result in less
wasted resources. Users can opt to increase this value based on
their environments.