Fixes #33577 - Only enable Foreman Puppet plugin with Foreman
If the user passed in --no-enable-foreman then the Foreman plugin should
not be enabled. The same is true for the CLI side of it.

The add-cli migration is changed to respect an already present option
that is false in order to properly test this change.

(cherry picked from commit ce89a30b315e729a45101a8707fcf88f9e9a39c8)