Refs #33864 - Disable registration on upgrades
In Foreman 3.1 the registration module requires templates where it
previously worked standalone.

In a Foreman or Katello scenario the registration module doesn't make
sense since the clients can connect directly to a Foreman, which makes
the indirection redundant. The Smart Proxy may also not have plain HTTP
enabled which then breaks things.

It is explicitly enabled by default in the foreman-proxy-content
scenario because there templates is enabled by default, which makes it
safe to enable registration.