Fixes #34590 - Enable only minimum apache modules
The puppetlabs-apache module installs Apache httpd with a set of default Apache modules, some of which are not necessary for our use case. This commit reduces that set of Apache modules to a minimum, by setting `apache::default_mods: false` in `config/foreman.hiera/common.yaml`. The user can enable additional modules if desired by overriding `apache::default_mods` in custom-hiera.yaml with a list of Apache modules. For more information, see documentation for puppetlabs-apache.