Fixes #35524 - Use a patched puppetlabs-apache for SSO
Without this Single Sign On using Keycloak doesn't work. The Apache
module needs certain other modules enabled. On EL8 it also needs to
enable a DNF module to install the package.

The fixes are upstream in 8.0.0 but various other modules don't allow
that. Using a patched module is the smallest impact.