Fixes #35629 - Default Apache to PROFILE=system ciphers
At least on EL8 it's possible to use PROFILE=system for SSLCipherSuite
and SSLProxyCipherSuite. This allows admins to configure the cipher
suite on a system level and it also means we don't have to keep our
cipher suite up to date.

Today SSLProxyCipherSuite is not yet an option, but Hiera will ignore
unknown keys. When the option becomes available, it will be set.