Ensure Ansible callback from the Ansible collection is used
This is only applied to the Foreman scenario because the parameter
always Red Hat always defaulted to the collection parameter since it was
introduced. It's mostly for Debian, but should be safe for RH too so no
facts are inspected.

(cherry picked from commit 416f45216493f9ad4d331a54d7cc92e42b8dbfd5)