Fixes #36561 - Run katello-certs-check on proxy cert generation
This is similar to how it runs on the regular installer. The advantage
of running this in foreman-proxy-certs-generate is that the user gets
the feedback earlier. So rather than running:

* foreman-proxy-certs-generate
* copy tar file to proxy
* run installer

And then finding out there was something wrong, the user gets the
feedback at the first step.

There could be a minor mismatch in version for n-1 version support, but
it's safe to assume the server will only be stricter but still
compatible. The check can be skipped if needed.