Refs #36573 - Drop puppet::server_foreman_url value if redundant
theforeman/puppetserver_foreman has started to default to copying the
value from foreman_proxy::foreman_base_url when unspecified.
theforeman/puppet has removed its default. This completes it by removing
the value if both answers are the same.

The biggest benefit is that the user only needs to care about a single
answer and no longer keep them in sync. This means we can stop thinking
about whether Puppet is enabled or not in all our user instructions.