Refs #36760 - Reset candlepin key- and truststore
This takes a 2 step approach where the cached password is removed during
migration (which ends up running during the RPM installation). Since we
don't want to break the system, the actual stores are only removed when
the installer runs. This will recreate them and update the matching
password files. Allowing a reset, even if theforeman/certs can't deal
with it.