Fixes #37686 - Stop puppetserver if java is too old
When upgrading from puppetserver 7 to puppetserver 8 it'll typically be
configured with /usr/bin/java and that usually points to Java 8.
Puppetserver 8 has started to require Java 11, so upgrading fails.

By properly stopping the service before running Puppet we know for sure
that it can be started safely again. It also helps free up CPU from the
constant restart loop, which reduces the slowdown.