Fixes #37693 - Add check for ipv6.disable=1 in /proc/cmdline
When a system has ipv6.disable=1 in /proc/cmdline then it doesn't have
::1 but some of our services (most notably, Redis) listen on it and
cause problems. Adding a check prevents users from installing in the
first place.

This check is based on definitions/checks/check_ipv6_disable.rb in
foreman_maintain. Biggest reason to add it here is to also guard it on
fresh installations.